Storge/multipart_uploader/requirements.txt
sanic == 0.7.0
requests
aiofiles
python-multipart
